Tree-shakeable Firestore
